pyproject.toml
[project]
name = "smolagents-with-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.37.32",
    "dependency-injector>=4.46.0",
    "fastapi>=0.115.12",
    "google-adk>=0.1.0",
    "litellm>=1.65.4.post1",
    "loguru>=0.7.3",
    "mcp[cli]>=1.6.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "ruff>=0.11.4",
    "smolagents[mcp]>=1.13.0",
    "typer>=0.15.2",
]
